Change the mechanism by which we manage scopes.  Instead of putting them in a hashmap keyed off of the context, put them into the contexts themselves.
This removes the need to close the scopes onDestroy, which caused any background threads to fail when trying to access providers.

It might be tempting to try to switch to weak references in the map instead, but you have to use both weak keys AND weak values, and weak values can get gc'd before the context does, resulting in non-singleton singletons.